\name{AssotesteR-internal}
\alias{print.assoctest}
\alias{print.rarecover}
\alias{print.stepup}
\alias{gdbr_IBS}
\alias{gdbr_wIBS}
\alias{kernel_IBS}
\alias{kernel_wIBS}
\alias{kernel_twowayx}
\title{Internal AssotesteR objects}
\description{Internal AssotesteR objects.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
