\name{AssotesteR-package}
\alias{AssotesteR}
\docType{package}
\title{
  Statistical Tests for Genetic Association Studies
}
\description{
  Statistical tests and methods for genetic association studies with emphasis on rare variants, binary (dichotomous) traits, and missing genotypes 
}
\details{
  \tabular{ll}{
    Package: \tab AssotesteR\cr
    Type: \tab Package\cr
    Version: \tab 0.1-9\cr
    Date: \tab 2013-01-23\cr
    Depends: \tab mvtnorm\cr
    License: \tab GPL (>= 3)\cr
    URL: \tab http://www.gastonsanchez.com/assotester \cr
    LazyLoad: \tab yes\cr
}
  The package contains several methods on testing for association between phenotype and genotype data with no covariates
}
\author{
  Author: Gaston Sanchez

  Maintainer: Gaston Sanchez <gaston.sanchez@berkeley.edu>
}
\references{
  \url{http://www.gastonsanchez.com/assotester}
}
\keyword{ package }


