\name{genodata}
\alias{genodata}
\docType{data}
\title{genodata}
\description{
  Simulated genotype data
}
\usage{data(genodata)}
\format{
  A data frame with 2000 observations on the following 21 variables.
  \describe{
    \item{\code{pheno}}{a numeric vector}
    \item{\code{snp1}}{a numeric vector}
    \item{\code{snp2}}{a numeric vector}
    \item{\code{snp3}}{a numeric vector}
    \item{\code{snp4}}{a numeric vector}
    \item{\code{snp5}}{a numeric vector}
    \item{\code{snp6}}{a numeric vector}
    \item{\code{snp7}}{a numeric vector}
    \item{\code{snp8}}{a numeric vector}
    \item{\code{snp9}}{a numeric vector}
    \item{\code{snp10}}{a numeric vector}
    \item{\code{snp11}}{a numeric vector}
    \item{\code{snp12}}{a numeric vector}
    \item{\code{snp13}}{a numeric vector}
    \item{\code{snp14}}{a numeric vector}
    \item{\code{snp15}}{a numeric vector}
    \item{\code{snp16}}{a numeric vector}
    \item{\code{snp17}}{a numeric vector}
    \item{\code{snp18}}{a numeric vector}
    \item{\code{snp19}}{a numeric vector}
    \item{\code{snp20}}{a numeric vector}
  }
}
\details{
  The first column is the phenotype (1000 cases coded as 1, 1000 controls coded as 0)
}
\examples{
  data(genodata)
}
\keyword{datasets}
